Move barrier attachment logic to AbstractNetworkWaySegmentWorldObject and attach stairs
